User Enter X or Z
created the general logic of what happens when a user enters X or Z. the while loop is true until the user enters X. in the case of Z, the program checks if Z was the first item entered from the user, then parses the input following the z as a string to be used as a file name. with this file name the program will try to open said file (assuming it's within the same directory). Included appropriate error checking for opening the file as well as for error conditions that occur after the while loop (which reads from the file line by line).